+++ packages/ossa/src/components/notice-bar/index.tsx
@@ -1,12 +1,13 @@
-import React from "react";
+import React, { CSSProperties, useEffect, useRef, useState } from "react";
+import { createSelectorQuery, getSystemInfoSync } from "@tarojs/taro";
 import { View } from "@tarojs/components";
 import classNames from "classnames";
 //引入组件对应的 类型文件 .d.ts
 import { OsNoticeBarProps } from "../../../types/index";
 import OsIcon from "../icon";
 
-function getStyleObj(props: OsNoticeBarProps) {
-  const _styleObj = {};
+function getStyleObj(props: OsNoticeBarProps): CSSProperties {
+  const _styleObj: CSSProperties = {};
   return _styleObj;
 }
 
@@ -18,10 +19,45 @@ function getClassObject(props: OsNoticeBarProps) {
   return _classObject;
 }
 
+async function getElementWidth(selector: string): Promise<number> {
+  return new Promise((resolve) => {
+    const query = createSelectorQuery();
+    query.select(selector).boundingClientRect();
+    query.exec(function (res) {
+      resolve(res?.[0]?.width || 0);
+    });
+  });
+}
+let uuid: number = 0;
+function getUUID() {
+  return ++uuid;
+}
 export default function Index(props: OsNoticeBarProps) {
   const rootClassName = "ossa-notice-bar"; //组件
   const classObject = getClassObject(props); //组件修饰
   const styleObject = Object.assign(getStyleObj(props), props.customStyle);
+  const [duration, setDuration] = useState(0);
+  const id = useRef<number>(getUUID());
+  const { speed = 16 } = props;
+
+  useEffect(() => {
+    const frameId = requestAnimationFrame(() => {
+      getElementWidth(`#marqueeContainer_${id.current}`).then((width) => {
+        if (width === 0) {
+          width = getSystemInfoSync().windowWidth;
+        }
+        setDuration(Number((width / speed).toFixed(2)));
+      });
+    });
+    return () => {
+      cancelAnimationFrame(frameId);
+    };
+  }, [speed]);
+
+  const marqueeStyle: CSSProperties = {
+    animationDuration: `${duration}s`,
+    WebkitAnimationDuration: `${duration}s`,
+  };
 
   return (
     <View
@@ -39,9 +75,11 @@ export default function Index(props: OsNoticeBarProps) {
       ) : null}
       <View className='ossa-notice-bar__content-wrapper'>
         <View
+          id={`marqueeContainer_${id.current}`}
           className={classNames({
             ["ossa-notice-bar__content"]: true,
           })}
+          style={marqueeStyle}
         >
           {props.children}
         </View>