+++ packages/ossa-doc/CHANGELOG.md
@@ -1,5 +1,11 @@
 # ossa-doc
 
+## 0.0.6
+
+### Patch Changes
+
+- [#78](https://github.com/NeteaseYanxuan/OSSA/pull/78) [`1039034`](https://github.com/NeteaseYanxuan/OSSA/commit/10390343772e0edacdf165f27866323b46dc2241) Thanks [@kiner-tang](https://github.com/kiner-tang)! - feat(noticeBar): 通知栏自定义滚动时长支持
+
 ## 0.0.5
 
 ### Patch Changes